MonoItemStack can now operate on the real item
Previously, if you created a MonoItemStack it would create a clone of
the item. If it was in an inventory or something, you had to replace the
existing one with the cloned one. This was because MonoItemStack could
only operate on CraftItemStacks. The new MonoItemStack can operate on
bukkit ItemStacks as well as CraftItemStacks so no clone is needed
anymore